third_party/blink/web_tests/virtual/omt-worker-fetch/http/tests/security/cors-rfc1918/addressspace-serviceworker-basic-expected.txt

This is a testharness.js-based test.
FAIL http://localhost:8000 = 'local' assert_equals: origin expected "http://localhost:8000" but got "ReferenceError"
FAIL http://127.0.0.1:8000 = 'local' assert_equals: origin expected "http://127.0.0.1:8000" but got "ReferenceError"
FAIL https://example.test:8443 = 'private' assert_equals: addressSpace expected "private" but got "public"
Harness: the test ran to completion.
